Fix missing collapse on trace detail
