+++ src/Aspire.Dashboard/Model/Otlp/SpanWaterfallViewModel.cs
@@ -113,4 +113,82 @@ public sealed class SpanWaterfallViewModel
             child.UpdateHidden(isParentCollapsed || IsCollapsed);
         }
     }
+
+    private readonly record struct SpanWaterfallViewModelState(SpanWaterfallViewModel? Parent, int Depth, bool Hidden);
+
+    public sealed record TraceDetailState(IEnumerable<IOutgoingPeerResolver> OutgoingPeerResolvers, List<string> CollapsedSpanIds);
+
+    public static List<SpanWaterfallViewModel> Create(OtlpTrace trace, TraceDetailState state)
+    {
+        var orderedSpans = new List<SpanWaterfallViewModel>();
+
+        TraceHelpers.VisitSpans(trace, (OtlpSpan span, SpanWaterfallViewModelState s) =>
+        {
+            var viewModel = CreateViewModel(span, s.Depth, s.Hidden, state);
+            orderedSpans.Add(viewModel);
+
+            s.Parent?.Children.Add(viewModel);
+
+            return s with { Parent = viewModel, Depth = s.Depth + 1, Hidden = viewModel.IsHidden || viewModel.IsCollapsed };
+        }, new SpanWaterfallViewModelState(Parent: null, Depth: 1, Hidden: false));
+
+        return orderedSpans;
+
+        static SpanWaterfallViewModel CreateViewModel(OtlpSpan span, int depth, bool hidden, TraceDetailState state)
+        {
+            var traceStart = span.Trace.FirstSpan.StartTime;
+            var relativeStart = span.StartTime - traceStart;
+            var rootDuration = span.Trace.Duration.TotalMilliseconds;
+
+            var leftOffset = relativeStart.TotalMilliseconds / rootDuration * 100;
+            var width = span.Duration.TotalMilliseconds / rootDuration * 100;
+
+            // Figure out if the label is displayed to the left or right of the span.
+            // If the label position is based on whether more than half of the span is on the left or right side of the trace.
+            var labelIsRight = (relativeStart + span.Duration / 2) < (span.Trace.Duration / 2);
+
+            // A span may indicate a call to another service but the service isn't instrumented.
+            var hasPeerService = OtlpHelpers.GetPeerAddress(span.Attributes) != null;
+            var isUninstrumentedPeer = hasPeerService && span.Kind is OtlpSpanKind.Client or OtlpSpanKind.Producer && !span.GetChildSpans().Any();
+            var uninstrumentedPeer = isUninstrumentedPeer ? ResolveUninstrumentedPeerName(span, state.OutgoingPeerResolvers) : null;
+
+            var viewModel = new SpanWaterfallViewModel
+            {
+                Children = [],
+                Span = span,
+                LeftOffset = leftOffset,
+                Width = width,
+                Depth = depth,
+                LabelIsRight = labelIsRight,
+                UninstrumentedPeer = uninstrumentedPeer
+            };
+
+            // Restore hidden/collapsed state to new view model.
+            if (state.CollapsedSpanIds.Contains(span.SpanId))
+            {
+                viewModel.IsCollapsed = true;
+            }
+            if (hidden)
+            {
+                viewModel.IsHidden = true;
+            }
+
+            return viewModel;
+        }
+    }
+
+    private static string? ResolveUninstrumentedPeerName(OtlpSpan span, IEnumerable<IOutgoingPeerResolver> outgoingPeerResolvers)
+    {
+        // Attempt to resolve uninstrumented peer to a friendly name from the span.
+        foreach (var resolver in outgoingPeerResolvers)
+        {
+            if (resolver.TryResolvePeerName(span.Attributes, out var name))
+            {
+                return name;
+            }
+        }
+
+        // Fallback to the peer address.
+        return OtlpHelpers.GetPeerAddress(span.Attributes);
+    }
 }
